fix: refactor author-head
