+++ packages/author-head/author-head-container.styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  wrapper: {
+    alignItems: "center",
+    backgroundColor: "transparent"
+  },
+  container: {
+    width: "100%",
+    alignItems: "center",
+    flexDirection: "column",
+    paddingBottom: 40,
+    backgroundColor: "#F9F8F3"
+  }
+});
+
+export default styles;
